Correct bank_transactions Schema

+++ dev/Setup/Sql/bank_transactions_schema.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS `bank_transactions` (
  `bank_transaction_id` int(11) NOT NULL AUTO_INCREMENT,
  `transaction_date` date NOT NULL,
  `description` varchar(200) NOT NULL,
- `amount` int(11) DEFAULT NULL,
+ `amount` decimal(8,2) DEFAULT NULL,
  `bank_id` int(11) NOT NULL,
  `account_id` int(11) NULL,
  PRIMARY KEY (`bank_transaction_id`)
